+++ code/python/src/utility/gnomonic_projection.py
@@ -96,7 +96,7 @@ def gnomonic_projection(theta, phi, theta_0, phi_0):
     # get cos_c's zero element index
     zeros_index = cos_c == 0
     if np.any(zeros_index):
-        cos_c[zeros_index] = np.finfo(np.float).eps
+        cos_c[zeros_index] = np.finfo(float).eps
 
     x = np.cos(phi) * np.sin(theta - theta_0) / cos_c
     y = (np.cos(phi_0) * np.sin(phi) - np.sin(phi_0) * np.cos(phi) * np.cos(theta - theta_0)) / cos_c
@@ -128,7 +128,7 @@ def reverse_gnomonic_projection(x, y, lambda_0, phi_1):
     # get rho's zero element index
     zeros_index = rho == 0
     if np.any(zeros_index):
-        rho[zeros_index] = np.finfo(np.float).eps
+        rho[zeros_index] = np.finfo(float).eps
 
     c = np.arctan2(rho, 1)
     phi_ = np.arcsin(np.cos(c) * np.sin(phi_1) + (y * np.sin(c) * np.cos(phi_1)) / rho)
@@ -187,11 +187,11 @@ def gnomonic2pixel(coord_gnom_x, coord_gnom_y,
     # TODO check add the padding whether necessary
     gnomonic2image_width_ratio = (tangent_image_width - 1.0) / (x_max - x_min + padding_size * 2.0)
     coord_pixel_x = (coord_gnom_x - x_min + padding_size) * gnomonic2image_width_ratio
-    coord_pixel_x = (coord_pixel_x + 0.5).astype(np.int)
+    coord_pixel_x = (coord_pixel_x + 0.5).astype(int)
 
     gnomonic2image_height_ratio = (tangent_image_height - 1.0) / (y_max - y_min + padding_size * 2.0)
     coord_pixel_y = -(coord_gnom_y - y_max - padding_size) * gnomonic2image_height_ratio
-    coord_pixel_y = (coord_pixel_y + 0.5).astype(np.int)
+    coord_pixel_y = (coord_pixel_y + 0.5).astype(int)
 
     return coord_pixel_x, coord_pixel_y
